Do support architecture specific setup
Sometimes different arch use different commands and that has to be
handled properly. For example s390x does not use network command to
setup the network. That's due to specifics of that architecture.

This patch contains basic support for this scenario as well as first
usage of that. The network command is not mandatory anymore if s390x
defined.